# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "ToDoList app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of ToDoList app"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://todolist-app-6cqj.onrender.com/",
      "description": "Production server"
    }
  ],
  "security": [],
  "tags": [
    {
      "name": "Task",
      "description": "Operations about task."
    }
  ],
  "paths": {
    "/task/{taskId}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get task by id",
        "operationId": "getTaskByID",
        "description": "Get task by id",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get task by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found task with id 65ca67e7ae7f10c88b598384!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/task"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "patch": {
        "tags": [
          "Tasks"
        ],
        "summary": "Update task",
        "operationId": "patchTask",
        "description": "Update a task with the provided payload. Only fields in the request body will be updated.",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            },
            "description": "ID of the task to update"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "task": {
                    "description": "Title of task",
                    "type": "string",
                    "example": "Water"
                  },
                  "description": {
                    "description": "The description of task",
                    "type": "string",
                    "example": "Water the plants in the house"
                  },
                  "columnName": {
                    "description": "Column to which this task belong",
                    "type": "string",
                    "enum": [
                      "ToDo",
                      "In Progress",
                      "Done"
                    ],
                    "default": "ToDo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated a task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated a task!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Delete task",
        "operationId": "deleteTask",
        "description": "Delete a task by id",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            },
            "description": "ID of the task to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully",
            "$ref": "#/components/responses/204"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/task": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get all tasks",
        "operationId": "getAllTasks",
        "description": "Get a list of all tasks",
        "responses": {
          "200": {
            "description": "Successfully get all tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found tasks!"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/task"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create task",
        "operationId": "createTask",
        "description": "Create a new task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "description": "Title of task",
                    "type": "string",
                    "example": "Water"
                  },
                  "description": {
                    "description": "The description of task",
                    "type": "string",
                    "example": "Water the plants in the house"
                  },
                  "columnName": {
                    "description": "Column to which this task belong",
                    "type": "string",
                    "enum": [
                      "ToDo",
                      "In Progress",
                      "Done"
                    ],
                    "default": "ToDo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a task!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "task": {
        "type": "object",
        "properties": {
          "title": {
            "description": "Title of task",
            "type": "string",
            "example": "Water"
          },
          "description": {
            "description": "The description of task",
            "type": "string",
            "example": "Water the pants in the house"
          },
          "columnName": {
            "description": "To which column does this task belong",
            "type": "string",
            "enum": [
              "ToDo",
              "In Progress",
              "Done"
            ],
            "default": "ToDo"
          }
        }
      }
    },
    "responses": {
      "204": {
        "description": "Contact successfully deleted",
        "content": {}
      },
      "400": {
        "description": "Invalid taskId",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Invalid taskId"
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Route not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Route not found"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Route not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}